chapter_07/requirements.txt
streamlit==1.40.2
pandas==2.2.2
plotly==5.23.0
humanize==4.10.0
toml==0.10.2
google-cloud-bigquery==3.25.0
google-cloud-bigquery-storage==2.26.0
db-dtypes==1.3.0